ProductosTest y DTOs

+++ backend/ms-productos/src/test/java/isi/dan/ms_productos/modelo/ProductoTest.java
@@ -2,8 +2,6 @@ package isi.dan.ms_productos.modelo;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.math.BigDecimal;
-
 import org.junit.jupiter.api.Test;
 
 public class ProductoTest {
@@ -23,7 +21,7 @@ public class ProductoTest {
             assertEquals("Producto " + categoria.name(), producto.getNombre());
             assertEquals(categoria, producto.getCategoria());
             assertEquals(10, producto.getStockActual());
-            assertEquals(new BigDecimal("99.99"), producto.getPrecio());
+            assertEquals(new Double("99.99"), producto.getPrecio());
         }
     }
